Move tests due to test refactor
