+++ tests/Aspire.Hosting.Tests/MongoDB/MongoDBFunctionalTests.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.Hosting.MongoDB;
 using Aspire.Hosting.Tests.Helpers;
 using Xunit;
 
@@ -29,24 +28,5 @@ public class MongoDBFunctionalTests
         var responseContent = await response.Content.ReadAsStringAsync();
 
         Assert.True(response.IsSuccessStatusCode, responseContent);
-    }
-
-    [Fact]
-    public void WithMongoExpressAddsContainer()
-    {
-        var builder = DistributedApplication.CreateBuilder();
-        builder.AddMongoDB("mongo").WithMongoExpress();
-
-        Assert.Single(builder.Resources.OfType<MongoExpressContainerResource>());
-    }
-
-    [Fact]
-    public void WithMongoExpressOnMultipleResources()
-    {
-        var builder = DistributedApplication.CreateBuilder();
-        builder.AddMongoDB("mongo").WithMongoExpress();
-        builder.AddMongoDB("mongo2").WithMongoExpress();
-
-        Assert.Equal(2, builder.Resources.OfType<MongoExpressContainerResource>().Count());
     }
 }
